Move RomeStepConfig into its own file since it's so big.
